backend/requirements.txt
Django==5.2.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
Pillow==12.1.1
psycopg2-binary==2.9.11
python-decouple==3.8
django-storages==1.14.6
boto3==1.42.59
gunicorn==25.1.0
whitenoise==6.12.0
